requirements.txt
Flask==0.8
Jinja2==2.6
Werkzeug==0.8.1
blinker==1.2
flatland==dev
argparse==1.2.1
CherryPy==3.2.2
Flask-Script==0.3.3
path.py==2.2.2
meld3==0.6.8
supervisor==3.0a12
simplejson==2.5.0
python-openid==2.2.5
Flask-OpenID==1.0.1
sqlitedict==1.0.8
